+++ AdapterUnitTests/AppLovinAdapterTests/AUTAppLovinAdapterTests.m
@@ -8,8 +8,6 @@
 #import <OCMock/OCMock.h>
 #import <XCTest/XCTest.h>
 
-#import "GADMAdapterAppLovinConstant.h"
-
 @interface AUTAppLovinAdapterTests : XCTestCase
 
 @end
@@ -59,7 +57,7 @@
   ALSdkInitializationConfigurationBuilder *builderMock =
       OCMClassMock([ALSdkInitializationConfigurationBuilder class]);
   OCMExpect([builderMock setMediationProvider:ALMediationProviderAdMob]);
-  OCMExpect([builderMock setPluginVersion:GADMAdapterAppLovinAdapterVersion]);
+  OCMExpect([builderMock setPluginVersion:GADMAdapterAppLovinConstant.adapterVersion]);
   OCMStub(ClassMethod([configMock configurationWithSdkKey:testSdkKey builderBlock:OCMOCK_ANY]))
       .andDo(^(NSInvocation *invocation) {
         __unsafe_unretained void (^block)(ALSdkInitializationConfigurationBuilder *builder);
@@ -76,7 +74,7 @@
       });
 
   AUTKMediationCredentials *credentials = [[AUTKMediationCredentials alloc] init];
-  credentials.settings = @{GADMAdapterAppLovinSDKKey : testSdkKey};
+  credentials.settings = @{GADMAdapterAppLovinConstant.sdkKey : testSdkKey};
   AUTKWaitAndAssertAdapterSetUpWithCredentials([GADMediationAdapterAppLovin class], credentials);
   OCMVerifyAll(appLovinSdkMock);
   OCMVerifyAll(builderMock);
@@ -97,7 +95,7 @@
   ALSdkInitializationConfigurationBuilder *builderMock =
       OCMClassMock([ALSdkInitializationConfigurationBuilder class]);
   OCMExpect([builderMock setMediationProvider:ALMediationProviderAdMob]);
-  OCMExpect([builderMock setPluginVersion:GADMAdapterAppLovinAdapterVersion]);
+  OCMExpect([builderMock setPluginVersion:GADMAdapterAppLovinConstant.adapterVersion]);
   OCMStub(ClassMethod([configMock configurationWithSdkKey:[OCMArg checkWithBlock:^BOOL(id obj) {
                                     return [testSdkKey1 isEqualToString:obj] ||
                                            [testSdkKey2 isEqualToString:obj];
@@ -118,10 +116,10 @@
       });
 
   AUTKMediationCredentials *credentials1 = [[AUTKMediationCredentials alloc] init];
-  credentials1.settings = @{GADMAdapterAppLovinSDKKey : testSdkKey1};
+  credentials1.settings = @{GADMAdapterAppLovinConstant.sdkKey : testSdkKey1};
   AUTKWaitAndAssertAdapterSetUpWithCredentials([GADMediationAdapterAppLovin class], credentials1);
   AUTKMediationCredentials *credentials2 = [[AUTKMediationCredentials alloc] init];
-  credentials2.settings = @{GADMAdapterAppLovinSDKKey : testSdkKey2};
+  credentials2.settings = @{GADMAdapterAppLovinConstant.sdkKey : testSdkKey2};
   AUTKWaitAndAssertAdapterSetUpWithCredentials([GADMediationAdapterAppLovin class], credentials2);
 
   OCMVerify(times(2), [appLovinSdkMock initializeWithConfiguration:OCMOCK_ANY
@@ -132,7 +130,7 @@
   GADMobileAds.sharedInstance.requestConfiguration.tagForChildDirectedTreatment = @YES;
   AUTKMediationCredentials *credentials = [[AUTKMediationCredentials alloc] init];
 
-  NSError *expectedError = [[NSError alloc] initWithDomain:GADMAdapterAppLovinErrorDomain
+  NSError *expectedError = [[NSError alloc] initWithDomain:GADMAdapterAppLovinConstant.errorDomain
                                                       code:GADMAdapterAppLovinErrorChildUser
                                                   userInfo:nil];
   AUTKWaitAndAssertAdapterSetUpFailureWithCredentials([GADMediationAdapterAppLovin class],
@@ -143,7 +141,7 @@
   GADMobileAds.sharedInstance.requestConfiguration.tagForUnderAgeOfConsent = @YES;
   AUTKMediationCredentials *credentials = [[AUTKMediationCredentials alloc] init];
 
-  NSError *expectedError = [[NSError alloc] initWithDomain:GADMAdapterAppLovinErrorDomain
+  NSError *expectedError = [[NSError alloc] initWithDomain:GADMAdapterAppLovinConstant.errorDomain
                                                       code:GADMAdapterAppLovinErrorChildUser
                                                   userInfo:nil];
   AUTKWaitAndAssertAdapterSetUpFailureWithCredentials([GADMediationAdapterAppLovin class],
@@ -155,7 +153,7 @@
       GADAgeRestrictedTreatmentChild;
   AUTKMediationCredentials *credentials = [[AUTKMediationCredentials alloc] init];
 
-  NSError *expectedError = [[NSError alloc] initWithDomain:GADMAdapterAppLovinErrorDomain
+  NSError *expectedError = [[NSError alloc] initWithDomain:GADMAdapterAppLovinConstant.errorDomain
                                                       code:GADMAdapterAppLovinErrorChildUser
                                                   userInfo:nil];
   AUTKWaitAndAssertAdapterSetUpFailureWithCredentials([GADMediationAdapterAppLovin class],
@@ -164,17 +162,17 @@
 
 - (void)testSetUpFailureWithInvalidSdkKey {
   NSString *testSdkKey = @"notValid";
-  NSError *error = [[NSError alloc] initWithDomain:GADMAdapterAppLovinErrorDomain
+  NSError *error = [[NSError alloc] initWithDomain:GADMAdapterAppLovinConstant.errorDomain
                                               code:GADMAdapterAppLovinErrorMissingSDKKey
                                           userInfo:nil];
   AUTKMediationCredentials *credentials = [[AUTKMediationCredentials alloc] init];
-  credentials.settings = @{GADMAdapterAppLovinSDKKey : testSdkKey};
+  credentials.settings = @{GADMAdapterAppLovinConstant.sdkKey : testSdkKey};
   AUTKWaitAndAssertAdapterSetUpFailureWithCredentials([GADMediationAdapterAppLovin class],
                                                       credentials, error);
 }
 
 - (void)testSetUpFailureWithMissingSdkKey {
-  NSError *error = [[NSError alloc] initWithDomain:GADMAdapterAppLovinErrorDomain
+  NSError *error = [[NSError alloc] initWithDomain:GADMAdapterAppLovinConstant.errorDomain
                                               code:GADMAdapterAppLovinErrorMissingSDKKey
                                           userInfo:nil];
   AUTKMediationCredentials *credentials = [[AUTKMediationCredentials alloc] init];